Rename chassis controller package
