+++ decomposition/metav_ros_control/meta_chassis_controller/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>omni_wheel_controller</name>
+  <name>meta_chassis_controller</name>
   <version>0.0.0</version>
-  <description>A ros2-control compatible omni wheel controller</description>
+  <description>A ros2-control compatible chassis controller for Meta-Team robots.</description>
   <maintainer email="chiming2@illinois.edu">Chiming Ni</maintainer>
   <license>MIT</license>
