--- decomposition/metav_ros_control/omni_wheel_controller/test/omni_wheel_controller_preceeding_params.yaml
@@ -1,28 +1,0 @@
-# Copyright (c) 20224 Stogl Robotics Consulting UG (haftungsbeschränkt)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-#
-# Source of this file are templates in
-# [RosTeamWorkspace](https://github.com/StoglRobotics/ros_team_workspace) repository.
-#
-
-test_omni_wheel_controller:
-  ros__parameters:
-    joints:
-      - joint1
-
-    interface_name: acceleration
-
-    state_joints:
-      - joint1state
